Disable process owner check if config is readonly
When 'config_is_read_only' has been enabled it is possible that not only
the config.php file is not writable by the webserver, but also that it
isn't owned by the webserver's uid at all.

So when enabled and executed in CLI mode cron.php should not compare
the owner of cron.php with the process owner as it may lead to wrong
positive as well as wrong negative results.

Refers to discussion in #8282.

Signed-off-by: Roland Tapken <roland@bitarbeiter.net>